refactor: make role service less visible
